+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # 📝 Apodiktum Modules Changelog:
 
+## 🆕 Version 2.0.3
+### 📃 module updates:
+- admintools | added `@admin` tag for chats (check config)
+- autoreact | added `ignore_self` config
+- dnd | automatically removes status after given time, improved bot message
+
 ## 🆕 Version 2.0.2
 ### 📃 module updates:
 - apoinfo | fixed apoinfo `{upd}`